Added CommResponse Unit tests for interfaces
